Most of this I did yesterday.  ISP pulled a brainfart and switched a few thousand customers including me to a brocken ras.
I've forgotten. basically I walked through linux rh71, solaris 2.6 and winNT4
and fixed compile/runtime issues.  This should be the uucp lock contrib and
nanosleep fixes on Solaris.  Compile fixes to match on win32 and linux.
update Documentation and merge with 1.5 docs as best as possible.

The should be a solid base to sync into.  there will be minor bug fixes as we
go.  I'll make a relase of this on rxtx.org tonight.